requirements.txt
fastapi==0.112.2
uvicorn[standard]==0.30.6
python-dotenv==1.0.1
pydantic==2.9.2
jinja2==3.1.4
websockets==12.0
httpx==0.27.2
aiohttp==3.10.11
openai==1.51.2
google-generativeai==0.8.3
pathway==0.13.0
feedparser==6.0.10
pytest==8.3.3
python-multipart==0.0.12
